Initial commit of ACLS Workbench dashboard/melon. There are some modifications to the core files (Marked with @Lucas):
- system/application/config/local_settings.php (added active_cover)
- system/application/controllers/system.php (added hook for active_cover, ammended duplicate feature, added "join" feature)
- system/application/languages/en/content_lang.php (added ACLS Workbench language; NEEDS TO BE AUGMENTED)
- system/application/libraries/SendMail.php (added function for sending "join" emails)
- system/application/models/book_model.php (tweaked the join/duplicate flags slightly to be more user-friendly)

Other modifications were either part of the previous ACLS or are new files.

Each edit has been made in such a way that the existing setups should not be affected, aside from the one edit to [system/application/config/local_settings.php] - I will make a note in the Changes-to-config-files-over-time wiki file.